refactor: Verschieben der Activity-Klassen in ein separates Paket
- Erstellung eines neuen Pakets "com.example.munchkin.activity" für die Activity-Klassen.
- Verschieben aller *Activity-Klassen in das neue Paket.
