+++ app/src/main/java/com/example/munchkin/activity/MainActivity.java
@@ -1,8 +1,10 @@
-package com.example.munchkin;
+package com.example.munchkin.activity;
 
 import android.os.Bundle;
 import androidx.activity.EdgeToEdge;
 import androidx.appcompat.app.AppCompatActivity;
+
+import com.example.munchkin.R;
 import com.example.munchkin.view.MainView;
 
 public class MainActivity extends AppCompatActivity {
